ceph: compensate for dentry_unhash() calls in vfs_rmdir() and vfs_rename_dir()
The VFS unhashes the dentry prior to calling the ->rmdir and ->rename
(directory) inode operations.  This is contrary to what the ->d_prune
d_op aims to provide: fs notification prior to dentries being unhashed.
Because it is done under the directory i_mutex and is thus serialized
against ->lookup() and ->readdir(), we can safely compensate by clearing
the flag in those methods.

Signed-off-by: Sage Weil <sage@newdream.net>